Change discord ready event
